fix: solve the bugs
