+++ frontend/krishi-hal-frontend/package.json
@@ -11,6 +11,8 @@
   },
   "dependencies": {
     "@tailwindcss/vite": "^4.0.17",
+    "@types/image-to-base64": "^2.1.2",
+    "image-to-base64": "^2.2.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "tailwindcss": "^4.0.17"
