+++ frontend/krishi-hal-frontend/src/hooks/useAudioRecorder.tsx
@@ -1,0 +1,45 @@
+import { useState, useRef } from "react";
+
+export const useAudioRecorder = (onComplete: (buffer: AudioBuffer) => void) => {
+  const [isRecording, setIsRecording] = useState(false);
+  const mediaRecorder = useRef<MediaRecorder | null>(null);
+  const audioChunks = useRef<Blob[]>([]);
+
+  const startRecording = async () => {
+    try {
+      const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
+      mediaRecorder.current = new MediaRecorder(stream);
+      audioChunks.current = [];
+
+      mediaRecorder.current.ondataavailable = (event) => {
+        if (event.data.size > 0) {
+          audioChunks.current.push(event.data);
+        }
+      };
+
+      mediaRecorder.current.onstop = async () => {
+        const audioBlob = new Blob(audioChunks.current, { type: "audio/wav" });
+        const arrayBuffer = await audioBlob.arrayBuffer();
+        const audioBuffer = await new AudioContext().decodeAudioData(
+          arrayBuffer
+        );
+        onComplete(audioBuffer);
+        stream.getTracks().forEach((track) => track.stop());
+      };
+
+      mediaRecorder.current.start();
+      setIsRecording(true);
+    } catch (error) {
+      console.error("Error accessing microphone:", error);
+    }
+  };
+
+  const stopRecording = () => {
+    if (mediaRecorder.current?.state === "recording") {
+      mediaRecorder.current.stop();
+      setIsRecording(false);
+    }
+  };
+
+  return { startRecording, stopRecording, isRecording };
+};
